+++ src/components/FileUpload.js
@@ -1,6 +1,6 @@
 function FileUpload(props) {
     return ( 
-        <input type="file" name={props.name} onChange={props.handleFile} className="form-control" required/>
+        <input type="file" name={props.name} onChange={props.handleFile} accept="image/png, image/jpeg" className="form-control" required/>
     );
 }
