+++ src/index.css
@@ -11,3 +11,34 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+.error{
+  margin-left: 5px;
+  font-size: 15px;
+  color: red;
+}
+
+.form{
+  margin: 30px;
+}
+
+.slider-class{
+  width: 300px;
+  margin: 30px;
+}
+
+.label-class{
+  display: flex;
+  flex-direction: column;
+  font-size: 22px;
+  align-items: center;
+  justify-content: center;
+}
+
+.box-image{
+	width: 150px;
+	height: 100px;
+	-webkit-box-shadow: -1px -3px 5px -2px rgba(214,214,214,1);
+  -moz-box-shadow: -1px -3px 5px -2px rgba(214,214,214,1);
+  box-shadow: -1px -3px 5px -2px rgba(214,214,214,1);
+}
